+++ src/ray.rs
@@ -14,7 +14,7 @@ impl Ray {
         }
     }
 
-    pub fn point_at(&self, t: f64) -> Vec3 {
+    pub fn point_at(&self, t: f32) -> Vec3 {
         self.origin + (t * self.direction)
     }
 }
